mi_unk1.c 71% (diamondtrust_us)

+++ include/nitro/mi.h
@@ -16,7 +16,7 @@ void MI_func_0206d87c(u32, void *src, void *dst, u32 size, u32, u32);
 void MI_func_0206d934(u32);
 BOOL MI_IsDmaBusy(u32);
 
-void _MI_CpuFill(u8 value, void *ptr, u32 size);
+void _MI_CpuFill(u32 value, void *ptr, u32 size);
 void MI_CpuFill8(void *ptr, u8 value, u32 size);
 void MI_CpuCopy8(void *src, void *dst, u32 size);
 void MI_CpuCopy16(void *src, void *dst, u32 size);
